+++ wire5/src/commonMain/kotlin/mqtt/wire5/control/packet/format/variable/property/TopicAlias.kt
@@ -5,6 +5,6 @@ package mqtt.wire5.control.packet.format.variable.property
 import kotlinx.io.core.BytePacketBuilder
 import mqtt.wire.data.Type
 
-data class TopicAlias(val value: UShort) : Property(0x22, Type.TWO_BYTE_INTEGER) {
-    override fun write(bytePacketBuilder: BytePacketBuilder) = write(bytePacketBuilder, value)
+data class TopicAlias(val value: Int) : Property(0x22, Type.TWO_BYTE_INTEGER) {
+    override fun write(bytePacketBuilder: BytePacketBuilder) = write(bytePacketBuilder, value.toUShort())
 }
